Move Visualizer management/init to `PlayerData`
